requirements.txt
tensorflow>=2.10.0
tensorflow-datasets>=4.8.0
opencv-python>=4.7.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
tqdm>=4.64.0
Pillow>=9.3.0
